requirements.txt
ansible==12.0.0
ansible-core==2.19.2
asgiref==3.9.1
black==24.10.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==46.0.1
dj-database-url==2.3.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy-extensions==1.0.0
opentelemetry-api==1.36.0
opentelemetry-distro==0.57b0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-dbapi==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-psycopg2==0.57b0
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==6.32.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python-memcached==1.62
PyYAML==6.0.3
redis==7.0.1
requests==2.32.3
resolvelib==1.2.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.9.0
wrapt==1.17.3
zipp==3.23.0
